rename SettingsPanel > Settings
