+++ src/components/SettingsPanel/Settings.styles.js
